rename and removes junk maybe
