+++ src/robin/encoders/__init__.py
@@ -1,2 +1,2 @@
-from .census_encoder import TableEncoder, YXDataset, YZDataset, ZDataset
 from .table_encoder import TableEncoder
+from .table_datasets import TableEncoder, YXDataset, YZDataset, ZDataset
